Extract helpers from the `import-triples-for-form` file
This file was originally the entry point and a lot of functions were added over time. Most of the helpers don't really have anything to do with "importing triples" so extracting them to separate files makes things easier to maintain.

It can probably be restructured even more, but this is a good start already.
