+++ src/index.js
@@ -1,5 +1,12 @@
+export * from "./add-simple-form-value.js";
 export { default } from "./constraints.js";
 export * from "./constraints.js";
+export * from "./fields-for-form.js";
 export * from "./generators.js";
 export * from "./import-triples-for-form.js";
 export * from "./namespaces.js";
+export * from "./remove-simple-form-value.js";
+export * from "./remove-triples.js";
+export * from "./triples-for.js";
+export * from "./update-simple-form-value.js";
+export * from "./validation.js";
